refactor(RD): unify download dialogs into a single file
No need for multiple files since the subclasses are so small now
